refactor: use cs2-kit PawnOps and EffectManager
Action bodies shrink to PawnOps calls (teleport clearance, swap,
bury, noclip/freeze/godmode toggles, slap), team constants come from
CS2Kit::Sdk, and the plugin-local EffectManager is replaced by the
kit registry keyed by int ids - EffectId stays as the plugin''s enum.

+++ plugins/admin-system/src/Admin/Effects/EffectAction.cpp
@@ -15,7 +15,7 @@ void Run(int adminSlot, int targetSlot, const EffectToggle& effect)
     if (!ctx.Valid())
         return;
 
-    bool on = App().Effects.Toggle(targetSlot, effect.Id, [&] { return effect.Enable(ctx); });
+    bool on = App().Effects.Toggle(targetSlot, static_cast<int>(effect.Id), [&] { return effect.Enable(ctx); });
     Actions::Broadcast(ctx, on ? effect.OnKey : effect.OffKey);
 }
