+++ plugins/admin-system/src/Admin/Menu/AdminMenu_Control.cpp
@@ -45,7 +45,7 @@ std::shared_ptr<::CS2Kit::Menu::Menu> BuildControlMenu(int adminSlot)
     // Self-only Hide toggle sits at the top of the Control list before player picks.
     builder.AddToggle(
         tr.Get("action.hide", adminSlot), tr.Get("effectState.on", adminSlot), tr.Get("effectState.off", adminSlot),
-        [adminSlot](int) { return App().Effects.IsActive(adminSlot, EffectId::Hide); },
+        [adminSlot](int) { return App().Effects.IsActive(adminSlot, static_cast<int>(EffectId::Hide)); },
         [adminSlot](int) { Effects::Run(adminSlot, adminSlot, Effects::Hide); }, hasB);
 
     auto players = plrMgr.GetAllPlayers();
